tool/hostexec: split linux sysproc tests
