+++ tool/hostexec/hostexec_unix_test.go
@@ -56,14 +56,6 @@ func waitForProcessExit(
 	t.Fatalf("background child pid %d is still alive", pid)
 }
 
-func TestApplyParentDeathSignal(t *testing.T) {
-	applyParentDeathSignal(nil)
-
-	attr := &syscall.SysProcAttr{}
-	applyParentDeathSignal(attr)
-	require.Equal(t, syscall.SIGTERM, attr.Pdeathsig)
-}
-
 func TestPrepareCommands(t *testing.T) {
 	preparePipeCommand(nil)
 	preparePTYCommand(nil)
@@ -72,13 +64,11 @@ func TestPrepareCommands(t *testing.T) {
 	preparePipeCommand(pipeCmd)
 	require.NotNil(t, pipeCmd.SysProcAttr)
 	require.True(t, pipeCmd.SysProcAttr.Setpgid)
-	require.Equal(t, syscall.SIGTERM, pipeCmd.SysProcAttr.Pdeathsig)
 
 	ptyCmd := &exec.Cmd{}
 	preparePTYCommand(ptyCmd)
 	require.NotNil(t, ptyCmd.SysProcAttr)
 	require.False(t, ptyCmd.SysProcAttr.Setpgid)
-	require.Equal(t, syscall.SIGTERM, ptyCmd.SysProcAttr.Pdeathsig)
 }
 
 func TestCommandProcessGroupID(t *testing.T) {
